misc: reorganize package
Prepare the structure to add new modules.

From an AI service prespective, we have several modules we offer:
openai-api and summaries today. These modules are exposed through
FeastAPI apps and mounted if enabled.

Make each module more encapsulated by ahving everything within.

If / when several modules need to share some functionality we can move
it out. At this stage, however, the persistence and job scheduling is
tied to summaries, for example, so it makes sense to have them together.
