--- skynet/models/v1/document.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-
-class DocumentPayload(BaseModel):
-    text: str
