+++ skynet/modules/ttt/summaries/app.py
@@ -5,10 +5,11 @@ from fastapi_versionizer.versionizer import versionize
 from skynet.env import enable_metrics
 from skynet.logs import get_logger
 from skynet.modules.monitoring import PROMETHEUS_NAMESPACE, PROMETHEUS_SUMMARIES_SUBSYSTEM
-from skynet.modules.persistence import db
-from skynet.modules.ttt.jobs import start_monitoring_jobs
-from skynet.modules.ttt.summaries import initialize as initialize_summaries
-from skynet.routers.v1 import router as v1_router
+
+from .persistence import db
+from .jobs import start_monitoring_jobs
+from .processor import initialize as initialize_summaries
+from .v1.router import router as v1_router
 
 
 log = get_logger('skynet.summaries')
