+++ skynet/modules/ttt/summaries/jobs_test.py
@@ -1,20 +1,20 @@
-from typing import Iterator
 import pytest
 
+from typing import Iterator
 from unittest.mock import patch
+
 from skynet.logs import get_logger
 
-from skynet.models.v1.document import DocumentPayload
-from skynet.models.v1.job import JobType
-from skynet.modules.persistence import db
-from skynet.modules.ttt.jobs import PENDING_JOBS_KEY
+from skynet.modules.ttt.summaries.persistence import db
+from skynet.modules.ttt.summaries.jobs import PENDING_JOBS_KEY
+from skynet.modules.ttt.summaries.v1.models import DocumentPayload, JobType
 
 log = get_logger('skynet.jobs_test')
 
 
 @pytest.fixture(scope='module', autouse=True)
 def default_session_fixture() -> Iterator[None]:
-    with patch('skynet.modules.persistence.db.set'), patch('skynet.modules.persistence.db.rpush'):
+    with patch('skynet.modules.ttt.summaries.persistence.db.set'), patch('skynet.modules.ttt.summaries.persistence.db.rpush'):
         yield
 
 
@@ -23,9 +23,9 @@ class TestCreateJob:
     async def test_runs_job(self, mocker):
         '''Test that a job is run.'''
 
-        mocker.patch('skynet.modules.ttt.jobs.create_run_job_task'),
+        mocker.patch('skynet.modules.ttt.summaries.jobs.create_run_job_task'),
 
-        from skynet.modules.ttt.jobs import create_job, create_run_job_task
+        from skynet.modules.ttt.summaries.jobs import create_job, create_run_job_task
 
         job_id = await create_job(JobType.SUMMARY, DocumentPayload(text='test'))
 
@@ -36,10 +36,10 @@ class TestCreateJob:
     async def test_queues_job(self, mocker):
         '''Test that a job is queued and queue size metric is updated.'''
 
-        mocker.patch('skynet.modules.ttt.jobs.can_run_next_job', return_value=False)
-        mocker.patch('skynet.modules.ttt.jobs.update_summary_queue_metric')
+        mocker.patch('skynet.modules.ttt.summaries.jobs.can_run_next_job', return_value=False)
+        mocker.patch('skynet.modules.ttt.summaries.jobs.update_summary_queue_metric')
 
-        from skynet.modules.ttt.jobs import create_job, update_summary_queue_metric
+        from skynet.modules.ttt.summaries.jobs import create_job, update_summary_queue_metric
 
         job_id = await create_job(JobType.SUMMARY, DocumentPayload(text='test'))
