+++ skynet/utils.py
